refactor: dictionary edit out of modal
